[PM] Regenerated demo for pm__pm_wakeup_rtc__pic32cm_pl10_curiosity_nano.X
MH3C-446

+++ apps/pm/pm_wakeup_rtc/src/config/default/peripheral/supc/plib_supc.c
@@ -85,7 +85,7 @@ void SUPC_MVIOVDDIO2ConfigSet(MVIO_VDDIO2CFG_CONFIG mvioVddio2Config)
 {
     if ((SUPC_REGS->SUPC_MVIO & SUPC_MVIO_MODE_Msk) == SUPC_MVIO_MODE_DUAL)
     {
-        SUPC_REGS->SUPC_MVIO = mvioVddio2Config;
+        SUPC_REGS->SUPC_MVIO = (uint32_t)mvioVddio2Config;
     }
 }
